test: move bundled-fetch content-length regression script to a fixture
