+++ test/node-test/global-dispatcher-version.js
@@ -102,37 +102,12 @@ test('requiring undici does not break Node.js global fetch with a request-set Co
   // Node's bundled fetch (undici v6/v7) appends its computed content-length
   // even when the request already carries one, producing "13, 13".
   // See https://github.com/nodejs/undici/issues/5500
-  const script = `
-    require('./index.js')
-    const http = require('node:http')
-    const { once } = require('node:events')
+  const result = spawnSync(
+    process.execPath,
+    [join(__dirname, '../fixtures/global-fetch-content-length.js')],
+    { cwd, encoding: 'utf8' }
+  )
 
-    ;(async () => {
-      const server = http.createServer((req, res) => {
-        let length = 0
-        req.on('data', (chunk) => { length += chunk.length })
-        req.on('end', () => res.end(String(length)))
-      })
-      server.listen(0)
-      await once(server, 'listening')
-
-      const body = 'update=INSERT'
-      const headers = new Headers()
-      headers.append('Content-Type', 'application/x-www-form-urlencoded')
-      headers.append('Content-Length', String(body.length))
-
-      const url = 'http://127.0.0.1:' + server.address().port
-      const res = await fetch(url, { method: 'POST', headers, body })
-      process.stdout.write(await res.text())
-
-      server.close()
-    })().catch((err) => {
-      console.error(err?.cause?.stack || err?.stack || err)
-      process.exit(1)
-    })
-  `
-
-  const result = runNode(script)
   assert.strictEqual(result.status, 0, result.stderr)
   assert.strictEqual(result.stdout, '13')
 })
